requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
celery[redis]>=5.3
django-celery-beat>=2.5
redis>=5.0
psycopg2-binary>=2.9
pdfplumber>=0.10
PyPDF2>=3.0
openai>=1.0,<1.13
